refactor(clawdstrike-policy-event): extract common from edr/receipt/mod.rs
